Introduce Occupant class
This commit replaces string literals like "plain" and "water" with
instances of an Occupant class. I think a dedicated class is more
appropriate because occupants must match a whitelist.

Occupant is commonly used primitive, so the literal syntax should be
quite concise. Thus, I defined Occ as a shorthand for Occupant.

I think the biggest problem is that now there are two types of
deserialization or "wrapping" going on. Previously, the only wrapping
distinction was between Generation instances and 2D arrays of strings.
Now each occupant can be represented as an Occupant or a string. So the
Generation.wrap method is expected to handle two different cases:
deserializing a 2D array of strings that we read from a file, and
constructing a Generation out of a 2D array of Occupants. (We
constructed a 2D array of Occupants rather than constructing a
Generation directly because it was easier not to do all the hooking of
up coordinates required by Generation.)

So maybe we need two types of wrapping/unwrapping. I'm not sure yet.