Renderers: Use Generation interface
I previously wanted to achieve a strict independence between the
renderers and the creators/evolvers. I don't really remember why - maybe
because I was going to move the renderers to another language. But I
plan to keep them in Ruby for now, and they'll benefit from many of the
same utilities defined in Generation and Cell.