Extract Grid
This extracts most of the geometrical methods into a Grid class.
Previously they were split between Cell and Generation.

This extraction makes the division of responsibilities clearer. It also
allows us to remove Cell's awareness of Generation, which will make it
possible to instantiate a bunch of Cells before instantiating their
Generation.